Rename utils file for org unit and periods steps
