+++ cypress/support/page-utils/categoryOptionsStep.js
@@ -1,0 +1,8 @@
+import { dataTest } from "../utils";
+
+export function selectAllAttributesCategoryOptions() {
+    cy.get(
+        '[data-test="FormControlLabel-sync-all-attribute-category-options"] > :nth-child(2)'
+    ).click();
+    cy.get(dataTest("group-editor-assign-all")).click();
+}
