+++ cypress/support/page-utils/orgUnitStep.js
@@ -9,3 +9,9 @@ export function assertSelectedOrgUnit(assert) {
 
     assert(cy.get(".ou-root .label > div[style*='color: orange;'"));
 }
+
+export function selectOrgUnit(container, orgUnit) {
+    cy.get(container).selectInOrgUnitTree(orgUnit);
+
+    return this;
+}
