--- cypress/support/page-utils/selectOrgUnitStep.js
@@ -1,5 +1,0 @@
-export function selectOrgUnit(container, orgUnit) {
-    cy.get(container).selectInOrgUnitTree(orgUnit);
-
-    return this;
-}
